Move OwncloudSyncService to NewsReaderListActivity
