+++ News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/NewsReaderListFragment.java
@@ -21,28 +21,12 @@
 
 package de.luhmer.owncloudnewsreader;
 
-import android.accounts.Account;
-import android.accounts.AccountManager;
 import android.app.Activity;
-import android.app.ActivityManager;
-import android.content.ComponentName;
-import android.content.ContentResolver;
 import android.content.Context;
-import android.content.Intent;
-import android.content.ServiceConnection;
 import android.content.SharedPreferences;
-import android.graphics.Color;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.IBinder;
-import android.os.Looper;
-import android.os.Parcelable;
-import android.os.RemoteException;
 import android.preference.PreferenceManager;
-import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnCreateContextMenuListener;
@@ -50,27 +34,15 @@ import android.view.ViewGroup;
 import android.widget.ExpandableListView;
 import android.widget.ExpandableListView.OnChildClickListener;
 import android.widget.ImageView;
-import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
-
-import org.apache.http.client.HttpResponseException;
-import org.apache.http.conn.HttpHostConnectException;
 
 import butterknife.ButterKnife;
 import butterknife.InjectView;
-import de.luhmer.owncloudnewsreader.Constants.SYNC_TYPES;
 import de.luhmer.owncloudnewsreader.ListView.SubscriptionExpandableListAdapter;
-import de.luhmer.owncloudnewsreader.authentication.AccountGeneral;
 import de.luhmer.owncloudnewsreader.database.DatabaseConnectionOrm;
-import de.luhmer.owncloudnewsreader.helper.AidlException;
-import de.luhmer.owncloudnewsreader.helper.PostDelayHandler;
 import de.luhmer.owncloudnewsreader.helper.ThemeChooser;
 import de.luhmer.owncloudnewsreader.interfaces.ExpListTextClicked;
 import de.luhmer.owncloudnewsreader.model.FolderSubscribtionItem;
-import de.luhmer.owncloudnewsreader.services.IOwnCloudSyncService;
-import de.luhmer.owncloudnewsreader.services.IOwnCloudSyncServiceCallback;
-import de.luhmer.owncloudnewsreader.services.OwnCloudSyncService;
 
 /**
  * A list fragment representing a list of NewsReader. This fragment also
@@ -82,84 +54,7 @@ import de.luhmer.owncloudnewsreader.services.OwnCloudSyncService;
  * interface.
  */
 public class NewsReaderListFragment extends Fragment implements OnCreateContextMenuListener {
-
-	IOwnCloudSyncService _ownCloudSyncService;
-	private IOwnCloudSyncServiceCallback callback = new IOwnCloudSyncServiceCallback.Stub() {
-
-		@Override
-		public void throwException(AidlException ex) throws RemoteException {
-			HandleExceptionMessages(ex.getmException());
-		}
-
-		@Override
-		public void startedSync(String sync_type) throws RemoteException {
-			Handler refresh = new Handler(Looper.getMainLooper());
-			refresh.post(new Runnable() {
-				public void run() {
-                    ((NewsReaderListActivity)getActivity()).UpdateButtonLayout();;
-				}
-			});
-		}
-
-		@Override
-		public void finishedSync(String sync_type) throws RemoteException {
-			Handler refresh = new Handler(Looper.getMainLooper());
-			refresh.post(new Runnable() {
-				public void run() {
-                    ((NewsReaderListActivity) getActivity()).UpdateButtonLayout();
-				}
-			});
-
-			SYNC_TYPES st = SYNC_TYPES.valueOf(sync_type);
-
-			switch(st) {
-				case SYNC_TYPE__GET_API:
-					break;
-				case SYNC_TYPE__ITEM_STATES:
-					break;
-				case SYNC_TYPE__FOLDER:
-					break;
-				case SYNC_TYPE__FEEDS:
-					break;
-				case SYNC_TYPE__ITEMS:
-
-					Log.d(TAG, "finished sync");
-					refresh = new Handler(Looper.getMainLooper());
-					refresh.post(new Runnable() {
-						public void run() {
-                        ReloadAdapter();
-                        NewsReaderListActivity nlActivity = (NewsReaderListActivity) getActivity();
-						if (nlActivity != null) {
-                            nlActivity.UpdateItemList();
-
-                            nlActivity.UpdatePodcastView();
-                        }
-
-                        SharedPreferences mPrefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
-                        int newItemsCount = mPrefs.getInt(Constants.LAST_UPDATE_NEW_ITEMS_COUNT_STRING, 0);
-                        if(newItemsCount > 0) {
-							Snackbar snackbar = Snackbar.make(nlActivity.findViewById(R.id.coordinator_layout), getResources().getQuantityString(R.plurals.message_bar_new_articles_available,newItemsCount,newItemsCount), Snackbar.LENGTH_LONG);
-							snackbar.setAction(getString(R.string.message_bar_reload), mListener);
-							snackbar.setActionTextColor(getResources().getColor(R.color.accent_material_dark));
-							// Setting android:TextColor to #000 in the light theme results in black on black
-							// text on the Snackbar, set the text back to white,
-							// TODO: find a cleaner way to do this
-							TextView textView = (TextView) snackbar.getView().findViewById(android.support.design.R.id.snackbar_text);
-							textView.setTextColor(Color.WHITE);
-							snackbar.show();
-                        }
-						}
-					});
-					break;
-			}
-		}
-
-	};
-
-	/*
-	public IOwnCloudSyncServiceCallback getCallback() {
-		return callback;
-	}*/
+	protected static final String TAG = "NewsReaderListFragment";
 
     public void ListViewNotifyDataSetChanged()  {
         lvAdapter.NotifyDataSetChangedAsync();
@@ -168,18 +63,6 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
     public void ReloadAdapter() {
         lvAdapter.ReloadAdapterAsync();
     }
-
-
-
-
-
-	/**
-	 * The serialization (saved instance state) Bundle key representing the
-	 * activated item position. Only used on tablets.
-	 */
-	//private static final String STATE_ACTIVATED_POSITION = "activated_position";
-
-	protected static final String TAG = "NewsReaderListFragment";
 
 	/**
 	 * The fragment's current callback object, which is notified of list item
@@ -198,11 +81,6 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 	}
 
 	/**
-	 * The current activated item position. Only used on tablets.
-	 */
-	//private static int mActivatedPosition = ListView.INVALID_POSITION;
-
-	/**
 	 * A callback interface that all activities containing this fragment must
 	 * implement. This mechanism allows activities to be notified of item
 	 * selections.
@@ -211,28 +89,19 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 		/**
 		 * Callback for when an item has been selected.
 		 */
-		public void onChildItemClicked(long idFeed, Long optional_folder_id);
-		public void onTopItemClicked(long idFeed, boolean isFolder, Long optional_folder_id);
+		void onChildItemClicked(long idFeed, Long optional_folder_id);
+		void onTopItemClicked(long idFeed, boolean isFolder, Long optional_folder_id);
 	}
 
 
-	//SubscriptionExpandableListAdapter lvAdapter;
 	private SubscriptionExpandableListAdapter lvAdapter;
-	//ExpandableListView eListView;
+
     @InjectView(R.id.expandableListView) ExpandableListView eListView;
 	@InjectView(R.id.urlTextView) protected TextView urlTextView;
 	@InjectView(R.id.userTextView) protected TextView userTextView;
 	@InjectView(R.id.header_view) protected ViewGroup headerView;
 	@InjectView(R.id.header_logo) protected ImageView headerLogo;
 	@InjectView(R.id.header_logo_progress) protected View headerLogoProgress;
-	//public static IReader _Reader = null;  //AsyncTask_GetGReaderTags asyncTask_GetUnreadFeeds = null;
-
-	//public static String username;
-	//public static String password;
-	//AsyncUpdateFinished asyncUpdateFinished;
-	ServiceConnection mConnection = null;
-
-    private View.OnClickListener mListener = null;
 
 	/**
 	 * Mandatory empty constructor for the fragment manager to instantiate the
@@ -240,186 +109,6 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 	 */
 	public NewsReaderListFragment() {
 	}
-
-	@Override
-	public void onCreate(Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
-
-		mListener = new View.OnClickListener()
-		{
-			@Override
-			public void onClick(View view)
-			{
-				//Toast.makeText(getActivity(), "button 1 pressed", 3000).show();
-
-				NewsReaderDetailFragment ndf = ((NewsReaderDetailFragment) getActivity().getSupportFragmentManager().findFragmentById(R.id.content_frame));
-				if(ndf != null) {
-					//ndf.reloadAdapterFromScratch();
-					ndf.UpdateCurrentRssView(getActivity(), true);
-				}
-			}
-		};
-	}
-
-
-
-	@Override
-	public void onStart() {
-		Intent serviceIntent = new Intent(getActivity(), OwnCloudSyncService.class);
-		mConnection = generateServiceConnection();
-        if(!isMyServiceRunning(OwnCloudSyncService.class)) {
-            getActivity().startService(serviceIntent);
-        }
-        getActivity().bindService(serviceIntent, mConnection, Context.BIND_AUTO_CREATE);
-		super.onStart();
-	}
-
-    private boolean isMyServiceRunning(Class<?> serviceClass) {
-        ActivityManager manager = (ActivityManager) getActivity().getSystemService(Context.ACTIVITY_SERVICE);
-        for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
-            if (serviceClass.getName().equals(service.service.getClassName())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-	@Override
-	public void onStop() {
-		if(_ownCloudSyncService != null) {
-			try {
-                _ownCloudSyncService.unregisterCallback(callback);
-			} catch (RemoteException e) {
-				e.printStackTrace();
-			}
-		}
-		getActivity().unbindService(mConnection);
-		mConnection = null;
-		super.onStop();
-	}
-
-	/* (non-Javadoc)
-	 * @see android.support.v4.app.Fragment#onResume()
-	 */
-	/*
-	@Override
-	public void onResume() {
-		try {
-			if(mActivatedPosition != ListView.INVALID_POSITION)
-				geteListView().getRefreshableView().setSelection(mActivatedPosition);
-		} catch(Exception ex) {
-			ex.printStackTrace();
-		}
-		super.onResume();
-	}
-	*/
-
-
-
-	/* (non-Javadoc)
-	 * @see android.support.v4.app.Fragment#onPause()
-	 */
-	@Override
-	public void onPause() {
-		//mActivatedPosition = geteListView().getRefreshableView().getFirstVisiblePosition();
-		super.onPause();
-	}
-
-	private ServiceConnection generateServiceConnection() {
-		return new ServiceConnection() {
-
-	    	@Override
-	    	public void onServiceConnected(ComponentName name, IBinder binder) {
-                _ownCloudSyncService = IOwnCloudSyncService.Stub.asInterface(binder);
-	    		try {
-                    _ownCloudSyncService.registerCallback(callback);
-
-	    			//Start auto sync if enabled
-	    			SharedPreferences mPrefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
-	    			if(mPrefs.getBoolean(SettingsActivity.CB_SYNCONSTARTUP_STRING, false))
-	    				StartSync();
-
-	    			if(getActivity() instanceof NewsReaderListActivity)
-	    				((NewsReaderListActivity) getActivity()).UpdateButtonLayout();
-	    		}
-	    		catch (Exception e) {
-	    			e.printStackTrace();
-	    		}
-	    	}
-
-	    	@Override
-	    	public void onServiceDisconnected(ComponentName name) {
-	    		try {
-                    _ownCloudSyncService.unregisterCallback(callback);
-	    		}
-	    		catch (Exception e) {
-	    			e.printStackTrace();
-	    		}
-	    	}
-		};
-    }
-
-
-	public void StartSync()
-	{
-		SharedPreferences mPrefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
-
-		if(mPrefs.getString(SettingsActivity.EDT_OWNCLOUDROOTPATH_STRING, null) == null)
-			NewsReaderListActivity.StartLoginFragment((FragmentActivity) getActivity());
-		else {
-			try {
-				if (!_ownCloudSyncService.isSyncRunning())
-				{
-					new PostDelayHandler(getActivity()).stopRunningPostDelayHandler();//Stop pending sync handler
-
-					//_ownCloadSyncService.startSync();
-
-					/*
-			         * Request the sync for the default account, authority, and
-			         * manual sync settings
-			         */
-					Bundle accBundle = new Bundle();
-					accBundle.putBoolean(ContentResolver.SYNC_EXTRAS_MANUAL, true);
-					AccountManager mAccountManager = AccountManager.get(getActivity());
-					Account[] accounts = mAccountManager.getAccounts();
-					for(Account acc : accounts)
-						if(acc.type.equals(AccountGeneral.ACCOUNT_TYPE))
-							ContentResolver.requestSync(acc, AccountGeneral.ACCOUNT_TYPE, accBundle);
-					//http://stackoverflow.com/questions/5253858/why-does-contentresolver-requestsync-not-trigger-a-sync
-				} else {
-                    ((NewsReaderListActivity) getActivity()).UpdateButtonLayout();
-                }
-			} catch (RemoteException e) {
-				e.printStackTrace();
-			}
-			//else
-	            //_Reader.attachToRunningTask(-10, getActivity(), onAsyncTask_GetVersionFinished);
-		}
-	}
-
-
-
-	private void HandleExceptionMessages(Exception ex) {
-		if(ex instanceof HttpHostConnectException)
-            ShowToastLong("Cannot connect to the Host !");
-        else if(ex instanceof HttpResponseException)
-        {
-            HttpResponseException responseException = (HttpResponseException) ex;
-            ShowToastLong(responseException.getLocalizedMessage());
-        }
-        else
-            ShowToastLong(ex.getLocalizedMessage());
-
-        ((NewsReaderListActivity) getActivity()).UpdateButtonLayout();
-	}
-
-
-
-    public void ShowToastLong(String message)
-    {
-        Toast.makeText(getActivity(), message, Toast.LENGTH_LONG).show();
-    }
-
 
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -452,7 +141,7 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 		headerView.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View v) {
-				StartSync();
+				((NewsReaderListActivity)getActivity()).startSync();
 			}
 		});
 
@@ -461,7 +150,6 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 
 		return view;
 	}
-
 
 	@Override
 	public void onAttach(Activity activity) {
@@ -491,19 +179,6 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 		}
     };
 
-	/*
-	@Override
-	public void onListItemClick(ListView listView, View view, int position,
-			long id) {
-		super.onListItemClick(listView, view, position, id);
-
-		// Notify the active callbacks interface (the activity, if the
-		// fragment is attached to one) that an item has been selected.
-
-		//mCallbacks.onItemSelected(DummyContent.ITEMS.get(position).id);
-	}*/
-
-
 	OnChildClickListener onChildClickListener = new OnChildClickListener() {
 
 		@Override
@@ -522,16 +197,4 @@ public class NewsReaderListFragment extends Fragment implements OnCreateContextM
 			return false;
 		}
 	};
-
-	/**
-	 * Turns on activate-on-click mode. When this mode is on, list items will be
-	 * given the 'activated' state when touched.
-	 */
-	public void setActivateOnItemClick(boolean activateOnItemClick) {
-		// When setting CHOICE_MODE_SINGLE, ListView will automatically
-		// give items the 'activated' state when touched.
-
-
-		eListView.setChoiceMode(activateOnItemClick ? ListView.CHOICE_MODE_SINGLE : ListView.CHOICE_MODE_NONE);
-	}
 }
