+++ News-Android-App/src/main/java/de/luhmer/owncloudnewsreader/PodcastFragmentActivity.java
@@ -156,7 +156,7 @@ public class PodcastFragmentActivity extends AppCompatActivity implements IPlayP
         super.onPause();
     }
 
-    private boolean isMyServiceRunning(Class<?> serviceClass) {
+    protected boolean isMyServiceRunning(Class<?> serviceClass) {
         ActivityManager manager = (ActivityManager) getSystemService(Context.ACTIVITY_SERVICE);
         for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
             if (serviceClass.getName().equals(service.service.getClassName())) {
